chatbot/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
pydantic==2.5.0
redis==5.0.1
psycopg2-binary==2.9.9
httpx==0.25.2
python-multipart==0.0.6
python-dotenv==1.0.0
aiofiles>=23.2.0
numpy>=1.24.0
asyncio
